pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "password-manager"
version = "0.0.3"
description = "Password Manager — backend (FastAPI)"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111,<1",
  "uvicorn[standard]>=0.30,<1",
  "pydantic>=2.7,<3",
  "platformdirs>=4.3.6",
  "argon2-cffi>=23.1.0",
  "pywebview>=4.4",
]

[project.optional-dependencies]
sqlcipher = [
  "sqlcipher3",
  "sqlcipher3-binary",
]

[tool.setuptools.packages.find]
where = [""]
include = ["pm*"]
